Refactoring to a Unix DriveInfo for basic info and OS specific DriveInfo instances for the GetDrives call
