+++ src/Common/src/Interop/Linux/Interop.Libraries.cs
@@ -5,6 +5,12 @@ internal static partial class Interop
 {
     private static partial class Libraries
     {
+        /// <summary>
+        /// We aren't OS X so we don't have a suffix
+        /// </summary>
+        internal const string INODE64SUFFIX = "";
+
         internal const string LibRt = "librt"; // POSIX Realtime Extensions library
+        internal const string libc = "libc";   // C runtime
     }
 }
