+++ src/Common/src/Interop/Unix/libc/Interop.mountpoints.cs
@@ -10,6 +10,8 @@ internal static partial class Interop
 {
     internal static partial class libc
     {
+        [DllImport("libc")]
+        static extern int printf(string data);
         /// <summary>
         /// Gets all the current mount points on the system
         /// </summary>
@@ -20,7 +22,7 @@ internal static partial class Interop
         /// Do NOT free this memory...this memory is allocated by the OS, which is responsible for it.
         /// This call could also block for a bit to wait for slow network drives.
         /// </remarks>
-        [DllImport(Interop.Libraries.libc, EntryPoint = "getmntinfo" + Interop.Libraries.INODE64SUFFIX)]
+        [DllImport(Interop.Libraries.libc, EntryPoint = "getmntinfo" + Interop.Libraries.INODE64SUFFIX, SetLastError = true)]
         internal static unsafe extern int getmntinfo(statfs** ppBuffer, int flags);
 
         /// <summary>
@@ -29,7 +31,7 @@ internal static partial class Interop
         /// <param name="path">The path to retrieve the statfs for</param>
         /// <param name="buffer">The output statfs struct describing the mount point</param>
         /// <returns>Returns 0 on success, -1 on failure</returns>
-        [DllImport(Interop.Libraries.libc, EntryPoint = "statfs" + Interop.Libraries.INODE64SUFFIX)]
+        [DllImport(Interop.Libraries.libc, EntryPoint = "statfs" + Interop.Libraries.INODE64SUFFIX, SetLastError = true)]
         private static unsafe extern int get_statfs(string path, statfs* buffer);
 
         /// <summary>
@@ -42,19 +44,18 @@ internal static partial class Interop
             statfs data = default(statfs);
             int result = get_statfs(name, &data);
             if (result < 0)
-                throw Interop.GetExceptionForIoErrno(Marshal.GetLastWin32Error());
+            {
+                int errno = Marshal.GetLastWin32Error();
+                if (errno == Interop.Errors.ENOENT)
+                    throw new System.IO.DriveNotFoundException(SR.Format(SR.IO_DriveNotFound_Drive, name)); // match Win32 exception
+                else
+                    throw Interop.GetExceptionForIoErrno(errno, isDirectory: true);
+            }
             else
+            {
+                printf("Should be here\r\n");
                 return data;
-        }
-
-        internal static unsafe String GetMountPointName(statfs data)
-        {
-            return Marshal.PtrToStringAnsi((IntPtr)data.f_mntonname);
-        }
-
-        internal static unsafe String GetMountPointFsType(statfs data)
-        {
-            return Marshal.PtrToStringAnsi((IntPtr)data.f_fstypename);
+            }
         }
     }
 }
